Moved Editable class into CinemachineScreenComposerGuides
